requirements.txt
transformers
torch
pandas
crewai
langchain-community
langchain-ollama
langchain-experimental
langchain-core
sqlalchemy
fastapi
uvicorn
pydantic
requests
python-dotenv